+++ test/test-dashes.sh
@@ -24,20 +24,20 @@ check obs exp
 echo "test-dashes.t02"
 gemini mendel_errors --columns "end,gt_types.Na-12877" test.dashes.db > obs
 echo "end	variant_id	gt_types.Na-12877	family_id	family_members	family_genotypes	samples	family_count	violation	violation_prob
-10671	1	1	CEPH1463	Na-12889(dad;unknown),Na-12890(mom;unknown),Na-12877(child;unknown)	G/G,G/G,G/C	Na-12877	1	plausible de novo	0.962
-28494	2	3	CEPH1463	Na-12889(dad;unknown),Na-12890(mom;unknown),Na-12877(child;unknown)	T/C,T/T,C/C	Na-12877	1	loss of heterozygosity	0.660
-28628	3	1	CEPH1463	Na-12889(dad;unknown),Na-12890(mom;unknown),Na-12877(child;unknown)	C/C,C/C,C/T	Na-12877	1	plausible de novo	0.989
-137123	4	0	CEPH1463	Na-12889(dad;unknown),Na-12890(mom;unknown),Na-12877(child;unknown)	T/T,G/G,T/T	Na-12877	1	uniparental disomy	0.881
-267560	5	1	CEPH1463	Na-12889(dad;unknown),Na-12890(mom;unknown),Na-12877(child;unknown)	C/C,C/C,CT/C	Na-12877	1	plausible de novo	0.896
-537541	6	1	CEPH1463	Na-12889(dad;unknown),Na-12890(mom;unknown),Na-12877(child;unknown)	T/T,T/T,T/TC	Na-12877	1	plausible de novo	0.729
-537970	7	1	CEPH1463	Na-12889(dad;unknown),Na-12890(mom;unknown),Na-12877(child;unknown)	C/C,C/C,C/T	Na-12877	1	plausible de novo	0.928
-540712	8	1	CEPH1463	Na-12889(dad;unknown),Na-12890(mom;unknown),Na-12877(child;unknown)	CGA/CGA,CGA/CGA,CGA/C	Na-12877	1	plausible de novo	0.814
-541052	9	1	CEPH1463	Na-12889(dad;unknown),Na-12890(mom;unknown),Na-12877(child;unknown)	T/T,T/T,T/C	Na-12877	1	plausible de novo	0.750
-547423	10	1	CEPH1463	Na-12889(dad;unknown),Na-12890(mom;unknown),Na-12877(child;unknown)	G/G,G/G,G/A	Na-12877	1	plausible de novo	0.746
-547519	11	1	CEPH1463	Na-12889(dad;unknown),Na-12890(mom;unknown),Na-12877(child;unknown)	G/G,G/G,G/T	Na-12877	1	plausible de novo	1.000
-547753	12	1	CEPH1463	Na-12889(dad;unknown),Na-12890(mom;unknown),Na-12877(child;unknown)	G/G,G/G,G/A	Na-12877	1	plausible de novo	0.754
-548246	13	1	CEPH1463	Na-12889(dad;unknown),Na-12890(mom;unknown),Na-12877(child;unknown)	T/T,T/T,T/TA	Na-12877	1	plausible de novo	0.925
-589086	14	3	CEPH1463	Na-12889(dad;unknown),Na-12890(mom;unknown),Na-12877(child;unknown)	G/G,GAGAA/GAGAA,G/G	Na-12877	1	uniparental disomy	0.940" > exp
+10671	1	1	CEPH1463	Na-12889(Na-12889;unknown;male),Na-12890(Na-12890;unknown;female),Na-12877(Na-12877;unknown;male)	G/G,G/G,G/C	Na-12877	1	plausible de novo	0.96228
+28494	2	3	CEPH1463	Na-12889(Na-12889;unknown;male),Na-12890(Na-12890;unknown;female),Na-12877(Na-12877;unknown;male)	T/C,T/T,C/C	Na-12877	1	loss of heterozygosity	0.65973
+28628	3	1	CEPH1463	Na-12889(Na-12889;unknown;male),Na-12890(Na-12890;unknown;female),Na-12877(Na-12877;unknown;male)	C/C,C/C,C/T	Na-12877	1	plausible de novo	0.98859
+137123	4	0	CEPH1463	Na-12889(Na-12889;unknown;male),Na-12890(Na-12890;unknown;female),Na-12877(Na-12877;unknown;male)	T/T,G/G,T/T	Na-12877	1	uniparental disomy	0.88112
+267560	5	1	CEPH1463	Na-12889(Na-12889;unknown;male),Na-12890(Na-12890;unknown;female),Na-12877(Na-12877;unknown;male)	C/C,C/C,CT/C	Na-12877	1	plausible de novo	0.89649
+537541	6	1	CEPH1463	Na-12889(Na-12889;unknown;male),Na-12890(Na-12890;unknown;female),Na-12877(Na-12877;unknown;male)	T/T,T/T,T/TC	Na-12877	1	plausible de novo	0.72864
+537970	7	1	CEPH1463	Na-12889(Na-12889;unknown;male),Na-12890(Na-12890;unknown;female),Na-12877(Na-12877;unknown;male)	C/C,C/C,C/T	Na-12877	1	plausible de novo	0.92824
+540712	8	1	CEPH1463	Na-12889(Na-12889;unknown;male),Na-12890(Na-12890;unknown;female),Na-12877(Na-12877;unknown;male)	CGA/CGA,CGA/CGA,CGA/C	Na-12877	1	plausible de novo	0.81372
+541052	9	1	CEPH1463	Na-12889(Na-12889;unknown;male),Na-12890(Na-12890;unknown;female),Na-12877(Na-12877;unknown;male)	T/T,T/T,T/C	Na-12877	1	plausible de novo	0.75000
+547423	10	1	CEPH1463	Na-12889(Na-12889;unknown;male),Na-12890(Na-12890;unknown;female),Na-12877(Na-12877;unknown;male)	G/G,G/G,G/A	Na-12877	1	plausible de novo	0.74611
+547519	11	1	CEPH1463	Na-12889(Na-12889;unknown;male),Na-12890(Na-12890;unknown;female),Na-12877(Na-12877;unknown;male)	G/G,G/G,G/T	Na-12877	1	plausible de novo	1.00000
+547753	12	1	CEPH1463	Na-12889(Na-12889;unknown;male),Na-12890(Na-12890;unknown;female),Na-12877(Na-12877;unknown;male)	G/G,G/G,G/A	Na-12877	1	plausible de novo	0.75392
+548246	13	1	CEPH1463	Na-12889(Na-12889;unknown;male),Na-12890(Na-12890;unknown;female),Na-12877(Na-12877;unknown;male)	T/T,T/T,T/TA	Na-12877	1	plausible de novo	0.92455
+589086	14	3	CEPH1463	Na-12889(Na-12889;unknown;male),Na-12890(Na-12890;unknown;female),Na-12877(Na-12877;unknown;male)	G/G,GAGAA/GAGAA,G/G	Na-12877	1	uniparental disomy	0.94024" > exp
 check obs exp
 rm obs exp